requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
authlib==1.3.0
python-dotenv==1.0.0
requests==2.31.0
pocketbase==0.1.0 # (Optional, though we will use SQLite for simplicity)
